Move `DemoError` into `examples/demo/demo_error.rs`.
